backend/telegram-callback/requirements.txt
psycopg2-binary==2.9.9
requests==2.31.0
PyMySQL==1.1.0
